refactor: Вынесены offers and distanations в отдельные модель-презентеры
