+++ src/model/destinations.js
@@ -1,0 +1,32 @@
+import Observer from "../view/utils/observer.js";
+
+export default class Destination extends Observer {
+  constructor() {
+    super();
+    this._destinations = [];
+  }
+
+  setDestinations(destinationsArray) {
+    this._destinations = destinationsArray.map((destination) => {
+      return this.adaptDestinationToClient(destination);
+    });
+  }
+
+  getDestinations() {
+    return this._destinations;
+  }
+
+  adaptDestinationToClient(destination) {
+    const adaptedDestination = Object.assign(
+        {},
+        destination,
+        {
+          photos: destination.pictures
+        }
+    );
+
+    delete adaptedDestination.pictures;
+
+    return adaptedDestination;
+  }
+}
